KAFKA-20812: Move Admin tests from KRaftClusterTest
